requirements.txt
# Core
pillow>=10.0.0
replicate>=0.25.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.26.0

# Face detection & similarity (InsightFace / ArcFace)
opencv-python>=4.10.0
insightface>=0.7.3
onnxruntime>=1.16.0

# Web interface
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.12
aiofiles>=24.1.0
